feat: remove legacy analysis page featureset
