+++ src/api/analysis/analysis.ts
@@ -6,7 +6,6 @@ import {
   GameNode,
   AnalyzedGame,
   MaiaEvaluation,
-  LegacyAnalyzedGame,
   PositionEvaluation,
   StockfishEvaluation,
   AnalysisTournamentGame,
@@ -111,306 +110,6 @@ export const getLichessGamePGN = async (id: string) => {
     },
   })
   return res.text()
-}
-
-export const getLegacyAnalyzedTournamentGame = async (
-  gameId = ['FkgYSri1'],
-) => {
-  const res = await fetch(
-    buildUrl(`analysis/analysis_list/${gameId.join('/')}`),
-  )
-
-  if (res.status === 401) {
-    throw new Error('Unauthorized')
-  }
-
-  const data = await res.json()
-  const id = data['id']
-  const termination = {
-    ...data['termination'],
-    condition: 'Normal',
-  }
-
-  const gameType = 'blitz'
-  const blackPlayer = data['black_player']
-  const whitePlayer = data['white_player']
-
-  const maiaEvaluations: { [model: string]: MoveMap[] } = {}
-  const stockfishEvaluations: MoveMap[] = data['stockfish_evals']
-
-  const positionEvaluations: { [model: string]: PositionEvaluation[] } = {}
-  const availableMoves: AvailableMoves[] = []
-
-  for (const model of data['maia_versions']) {
-    maiaEvaluations[model] = data['maia_evals'][model]
-    positionEvaluations[model] = Object.keys(data['maia_evals'][model]).map(
-      () => ({
-        trickiness: 1,
-        performance: 1,
-      }),
-    )
-  }
-
-  for (const position of data['move_maps']) {
-    const moves: AvailableMoves = {}
-    for (const move of position) {
-      const fromTo = move.move.join('')
-      const san = move['move_san']
-      const { check, fen } = move
-
-      moves[fromTo] = {
-        board: fen,
-        check,
-        san,
-        lastMove: move.move,
-      }
-    }
-    availableMoves.push(moves)
-  }
-
-  const gameStates = data['game_states']
-
-  const moves = gameStates.map((gameState: any) => {
-    const {
-      last_move: lastMove,
-      fen,
-      check,
-      last_move_san: san,
-      evaluations: maia_values,
-    } = gameState
-
-    return {
-      board: fen,
-      lastMove,
-      san,
-      check,
-      maia_values,
-    }
-  })
-
-  return {
-    id,
-    blackPlayer,
-    whitePlayer,
-    moves,
-    maiaEvaluations,
-    stockfishEvaluations,
-    availableMoves,
-    gameType,
-    termination,
-    positionEvaluations,
-  } as any as LegacyAnalyzedGame
-}
-
-export const getLegacyAnalyzedLichessGame = async (
-  id: string,
-  pgn: string,
-  maia_model = 'maia_kdd_1500',
-) => {
-  const res = await fetch(
-    buildUrl(
-      'analysis/analyze_user_game?' +
-        new URLSearchParams({
-          maia_model,
-        }),
-    ),
-    {
-      method: 'POST',
-      body: pgn,
-      headers: {
-        'Content-Type': 'text/plain',
-      },
-    },
-  )
-
-  if (res.status === 401) {
-    throw new Error('Unauthorized')
-  }
-
-  const data = await res.json()
-
-  const termination = {
-    ...data['termination'],
-    condition: 'Normal',
-  }
-
-  const gameType = 'blitz'
-  const blackPlayer = data['black_player']
-  const whitePlayer = data['white_player']
-
-  const maiaEvaluations: { [model: string]: MoveMap[] } = {}
-  const positionEvaluations: { [model: string]: PositionEvaluation[] } = {}
-  const availableMoves: AvailableMoves[] = []
-
-  for (const model of data['maia_versions']) {
-    maiaEvaluations[model] = data['maia_evals'][model]
-    positionEvaluations[model] = Object.keys(data['maia_evals'][model]).map(
-      () => ({
-        trickiness: 1,
-        performance: 1,
-      }),
-    )
-  }
-
-  for (const position of data['move_maps']) {
-    const moves: AvailableMoves = {}
-    for (const move of position) {
-      const fromTo = move.move.join('')
-      const san = move['move_san']
-      const { check, fen } = move
-
-      moves[fromTo] = {
-        board: fen,
-        check,
-        san,
-        lastMove: move.move,
-      }
-    }
-    availableMoves.push(moves)
-  }
-
-  const gameStates = data['game_states']
-
-  const moves = gameStates.map((gameState: any) => {
-    const {
-      last_move: lastMove,
-      fen,
-      check,
-      last_move_san: san,
-      evaluations: maia_values,
-    } = gameState
-
-    return {
-      board: fen,
-      lastMove,
-      san,
-      check,
-      maia_values,
-    }
-  })
-
-  return {
-    id,
-    blackPlayer,
-    whitePlayer,
-    moves,
-    maiaEvaluations,
-    availableMoves,
-    gameType,
-    termination,
-    positionEvaluations,
-    pgn,
-  } as LegacyAnalyzedGame
-}
-
-export const getLegacyAnalyzedUserGame = async (
-  id: string,
-  game_type: 'play' | 'hand' | 'brain',
-  maia_model = 'maia_kdd_1500',
-) => {
-  const res = await fetch(
-    buildUrl(
-      `analysis/user/analyze_user_maia_game/${id}?` +
-        new URLSearchParams({
-          game_type,
-          maia_model,
-        }),
-    ),
-    {
-      method: 'GET',
-      headers: {
-        'Content-Type': 'text/plain',
-      },
-    },
-  )
-
-  if (res.status === 401) {
-    throw new Error('Unauthorized')
-  }
-
-  const data = await res.json()
-
-  const termination = {
-    ...data['termination'],
-    condition: 'Normal',
-  }
-
-  const gameType = 'blitz'
-  const blackPlayer = data['black_player']
-  const whitePlayer = data['white_player']
-
-  const maiaPattern = /maia_kdd_1\d00/
-
-  if (blackPlayer.name && maiaPattern.test(blackPlayer.name)) {
-    blackPlayer.name = blackPlayer.name.replace('maia_kdd_', 'Maia ')
-  }
-
-  if (whitePlayer.name && maiaPattern.test(whitePlayer.name)) {
-    whitePlayer.name = whitePlayer.name.replace('maia_kdd_', 'Maia ')
-  }
-
-  const maiaEvaluations: { [model: string]: MoveMap[] } = {}
-  const positionEvaluations: { [model: string]: PositionEvaluation[] } = {}
-  const availableMoves: AvailableMoves[] = []
-
-  for (const model of data['maia_versions']) {
-    maiaEvaluations[model] = data['maia_evals'][model]
-    positionEvaluations[model] = Object.keys(data['maia_evals'][model]).map(
-      () => ({
-        trickiness: 1,
-        performance: 1,
-      }),
-    )
-  }
-
-  for (const position of data['move_maps']) {
-    const moves: AvailableMoves = {}
-    for (const move of position) {
-      const fromTo = move.move.join('')
-      const san = move['move_san']
-      const { check, fen } = move
-
-      moves[fromTo] = {
-        board: fen,
-        check,
-        san,
-        lastMove: move.move,
-      }
-    }
-    availableMoves.push(moves)
-  }
-
-  const gameStates = data['game_states']
-
-  const moves = gameStates.map((gameState: any) => {
-    const {
-      last_move: lastMove,
-      fen,
-      check,
-      last_move_san: san,
-      evaluations: maia_values,
-    } = gameState
-
-    return {
-      board: fen,
-      lastMove,
-      san,
-      check,
-      maia_values,
-    }
-  })
-
-  return {
-    id,
-    blackPlayer,
-    whitePlayer,
-    moves,
-    maiaEvaluations,
-    availableMoves,
-    gameType,
-    termination,
-    positionEvaluations,
-  } as LegacyAnalyzedGame
 }
 
 function convertMoveMapToStockfishEval(
