+++ src/hooks/useTrainingController/index.ts
@@ -1,2 +1,1 @@
 export * from './useTrainingController'
-export * from './useTrainingTreeController'
